refactor(react): move internal FeatureAppContainer/Loader to external files
